+++ hedera-node/hedera-config/src/main/java/com/hedera/node/config/data/TssConfig.java
@@ -19,7 +19,7 @@ public record TssConfig(
         @ConfigProperty(defaultValue = "10s") @NetworkProperty Duration crsUpdateContributionTime,
         @ConfigProperty(defaultValue = "5s") @NetworkProperty Duration crsFinalizationDelay,
         @ConfigProperty(defaultValue = "data/keys/tss") @NodeProperty String tssKeysPath,
-        @ConfigProperty(defaultValue = "1024") @NetworkProperty int initialCrsParties,
+        @ConfigProperty(defaultValue = "1024") @NetworkProperty short initialCrsParties,
         @ConfigProperty(defaultValue = "false") @NetworkProperty boolean hintsEnabled,
         @ConfigProperty(defaultValue = "false") @NetworkProperty boolean historyEnabled,
         // Must be true if enabling TSS while also using an override network,
